python.inode: correct type casting like python.device
Disregard older Matlab <= R2022a as they can use other methods, Python doesn't work as well
with those old Matlab versions anyway.

+++ +stdlib/+python/inode.m
@@ -1,12 +1,14 @@
 function i = inode(file)
 
 try
-  i = int64(py.os.stat(file).st_ino);
-  % int64 first is for Matlab <= R2022a
+  i = py.os.stat(file).st_ino;
+  % Matlab <= R2022a wants int64 before uint64, but this can make OverflowError on Windows within Python
+  % because on Windows st_dev is a 64-bit unsigned integer
 catch e
   pythonException(e)
   i = [];
 end
 
 i = uint64(i);
+
 end
